requirements.txt
asgiref==3.8.1
Django==4.2.11
djangorestframework==3.15.1
psycopg2-binary==2.9.9
python-json-logger==2.0.7
sqlparse==0.4.4
typing_extensions==4.11.0
boto3~=1.34.82
python-decouple==3.8
python-decouple==3.8
uWSGI==2.0.25
requests==2.32.3
pytz==2024.2
langgraph==0.2.62
langsmith==0.2.10
langchain-anthropic==0.3.1
faiss-cpu==1.9.0.post1

openai==1.59.9
langchain-openai==0.3.1
langchain==0.3.15
langchain-core==0.3.31
langchain-community==0.3.15